+++ parseconfig.go
@@ -17,12 +17,15 @@ import (
 // DjConfig is a Golang struct representation of mumbledj.gcfg file structure for parsing.
 type DjConfig struct {
 	General struct {
-		CommandPrefix     	string
-		SkipRatio         	float32
-		PlaylistSkipRatio 	float32
-		DefaultComment    	string
-		MaxSongDuration   	int
-		AutomaticShuffleOn  bool
+		CommandPrefix      string
+		SkipRatio          float32
+		PlaylistSkipRatio  float32
+		DefaultComment     string
+		MaxSongDuration    int
+		MaxSongPerPlaylist int
+		AutomaticShuffleOn bool
+		AnnounceNewTrack   bool
+		PlayerCommand      string
 	}
 	Cache struct {
 		Enabled     bool
@@ -36,14 +39,14 @@ type DjConfig struct {
 	}
 	Aliases struct {
 		AddAlias               string
-		AddAlias2              string
+		AddNextAlias           string
 		SkipAlias              string
 		SkipPlaylistAlias      string
 		AdminSkipAlias         string
 		AdminSkipPlaylistAlias string
 		HelpAlias              string
 		VolumeAlias            string
-		VolumeAlias2           string
+		VolumeAliasTwo         string
 		MoveAlias              string
 		ReloadAlias            string
 		ResetAlias             string
@@ -57,27 +60,34 @@ type DjConfig struct {
 		ShuffleAlias           string
 		ShuffleOnAlias         string
 		ShuffleOffAlias        string
+		ListSongsAlias         string
 	}
 	Permissions struct {
-		AdminsEnabled     	bool
-		Admins            	[]string
-		AdminAdd          	bool
-		AdminAddPlaylists 	bool
-		AdminSkip         	bool
-		AdminHelp         	bool
-		AdminVolume       	bool
-		AdminMove         	bool
-		AdminReload       	bool
-		AdminReset        	bool
-		AdminNumSongs     	bool
-		AdminNextSong     	bool
-		AdminCurrentSong  	bool
-		AdminSetComment   	bool
-		AdminNumCached    	bool
-		AdminCacheSize    	bool
-		AdminKill         	bool
-		AdminShuffle        bool
-		AdminShuffleToggle  bool
+		AdminsEnabled      bool
+		Admins             []string
+		AdminAdd           bool
+		AdminAddNext       bool
+		AdminAddPlaylists  bool
+		AdminSkip          bool
+		AdminHelp          bool
+		AdminVolume        bool
+		AdminMove          bool
+		AdminReload        bool
+		AdminReset         bool
+		AdminNumSongs      bool
+		AdminNextSong      bool
+		AdminCurrentSong   bool
+		AdminSetComment    bool
+		AdminNumCached     bool
+		AdminCacheSize     bool
+		AdminKill          bool
+		AdminShuffle       bool
+		AdminShuffleToggle bool
+		AdminListSongs     bool
+	}
+	ServiceKeys struct {
+		Youtube    string
+		SoundCloud string
 	}
 }