+++ strings.go
@@ -99,6 +99,16 @@ const PLAYLIST_ADDED_HTML = `
 	<b>%s</b> has added the playlist "%s" to the queue.
 `
 
+// Message shown to channel when a song is added to the queue by a user after the current song.
+const NEXT_SONG_ADDED_HTML = `
+	<b>%s</b> has added "%s" to the queue after the current song.
+`
+
+// Message shown to channel when a playlist is added to the queue by a user after the current song.
+const NEXT_PLAYLIST_ADDED_HTML = `
+	<b>%s</b> has added the playlist "%s" to the queue after the current song.
+`
+
 // Message shown to channel when a song has been skipped.
 const SONG_SKIPPED_HTML = `
 	The number of votes required for a skip has been met. <b>Skipping song!</b>
@@ -118,13 +128,18 @@ const HELP_HTML = `<br/>
 	<p><b>!skip</b> - Casts a vote to skip the current song</p>
 	<p> <b>!skipplaylist</b> - Casts a vote to skip over the current playlist.</p>
 	<p><b>!numsongs</b> - Shows how many songs are in queue.</p>
+	<p><b>!listsongs</b> - Lists the songs in queue.</p>
 	<p><b>!nextsong</b> - Shows the title and submitter of the next queue item if it exists.</p>
 	<p><b>!currentsong</b> - Shows the title and submitter of the song currently playing.</p>
 	<p style="-qt-paragraph-type:empty"><br/></p>
 	<p><b>Admin Commands:</b></p>
+	<p><b>!addnext</b> - Adds songs/playlists to queue after the current song.</p>
 	<p><b>!reset</b> - An admin command that resets the song queue. </p>
 	<p><b>!forceskip</b> - An admin command that forces a song skip. </p>
 	<p><b>!forceskipplaylist</b> - An admin command that forces a playlist skip. </p>
+	<p><b>!shuffle</b> - An admin command that shuffles the current queue. </p>
+	<p><b>!shuffleon</b> - An admin command that enables auto shuffling.</p>
+  	<p><b>!shuffleoff</b> - An admin command that disables auto shuffling.</p>
 	<p><b>!move </b>- Moves MumbleDJ into channel if it exists.</p>
 	<p><b>!reload</b> - Reloads mumbledj.gcfg configuration settings.</p>
 	<p><b>!setcomment</b> - Sets the comment for the bot.</p>
@@ -186,5 +201,10 @@ const CURRENT_SONG_HTML = `
 // Message shown to users when the currentsong command is issued when a song from a
 // playlist is playing.
 const CURRENT_SONG_PLAYLIST_HTML = `
-	The %s currently playing is "%s", added <b>%s</b> from the %s "%s".
+	The song currently playing is "%s", added <b>%s</b> from the playlist "%s".
+`
+
+// Message shown to user when the listsongs command is issued
+const SONG_LIST_HTML = `
+	<br>%d: "%s", added by <b>%s</b>.</br>
 `